requirements.txt
aiokafka==0.8.1
asr-python==0.5.0
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
flake8==6.0.0
flasgger==0.9.7.1
Flask==1.1.4
gunicorn==20.0.4
idna==3.4
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kafka-python==2.0.2
MarkupSafe==2.0.1
mccabe==0.7.0
mistune==3.0.1
msal==1.23.0
msal-extensions==1.0.0
packaging==23.1
portalocker==2.7.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==1.0.1

colorama==0.4.4
shellingham==1.4.0
pytest==6.2.4
selenium==4.11.2
chromedriver-py==114.0.5735.90
beautifulsoup4==4.12.2
opencv-python==4.8.0.74
scipy==1.10.1
aiohttp==3.8.5
cryptography==41.0.3

## The following requirements were added by pip freeze:
imageio==2.31.1
imutils==0.5.4
scikit-image==0.21.0
## The following requirements were added by pip freeze:
aiosignal==1.3.1
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
iniconfig==2.0.0
lazy_loader==0.3
multidict==6.0.4
networkx==3.1
numpy==1.25.2
outcome==1.2.0
Pillow==10.0.0
pluggy==0.13.1
py==1.11.0
PySocks==1.7.1
PyWavelets==1.4.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tifffile==2023.8.12
toml==0.10.2
trio==0.22.2
trio-websocket==0.10.3
wsproto==1.2.0
yarl==1.9.2
python-json-logger==2.0.7

fastapi~=0.111.0
dnspython~=2.6.1
httpcore~=1.0.5
httpx~=0.27.0
anyio~=4.3.0
pip~=24.0
wheel~=0.41.2
rich~=13.7.1
pygments~=2.18.0
mdurl~=0.1.2
typer~=0.3.2
orjson~=3.10.3
uvloop~=0.19.0
starlette~=0.37.2
pydantic~=2.7.1
websockets~=12.0
ujson~=5.10.0
uvicorn~=0.29.0
httptools~=0.6.1
watchfiles~=0.21.0
python-multipart~=0.0.9
setuptools~=68.2.0
redis~=5.0.4